refactor(kms): split service.rs into resource-grouped impl files
service.rs grew to 3008 lines after Phase I extraction. Move four
resource-grouped method clusters into sibling files via #[path]
modules with impl KmsService blocks:
- service_crypto.rs (821 lines): encrypt/decrypt/re_encrypt, all
  data-key/data-key-pair generators, sign/verify, generate/verify_mac,
  generate_random, derive_shared_secret, get_public_key.
- service_aliases.rs (193 lines): create/delete/update/list_alias.
- service_grants.rs (247 lines): create/list/list_retirable/revoke/retire_grant.
- service_custom_store.rs (499 lines): custom key store CRUD +
  connect/disconnect, key-material import lifecycle.

Methods exposed as pub(super) so service_tests.rs sibling can call them.
service.rs: 3008 -> 1310 lines (-56%).
