+++ crates/fakecloud-kms/src/service.rs
@@ -724,470 +724,6 @@ impl KmsService {
         ))
     }
 
-    fn encrypt(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let key_id = Self::require_key_id(&body)?;
-        let plaintext_b64 = body["Plaintext"].as_str().ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                "Plaintext is required",
-            )
-        })?;
-        let plaintext_bytes = decode_plaintext(plaintext_b64)?;
-
-        let resolved = self
-            .resolve_key_id_for(&req.account_id, &req.region, &key_id)
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "NotFoundException",
-                    format!("Key '{key_id}' does not exist"),
-                )
-            })?;
-
-        let accounts = self.state.read();
-        let empty = KmsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        let key = state.keys.get(&resolved).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "KMSInternalException",
-                "Key state became inconsistent",
-            )
-        })?;
-        if !key.enabled {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "DisabledException",
-                format!("Key '{}' is disabled", key.arn),
-            ));
-        }
-
-        let ciphertext_b64 = build_encrypt_ciphertext(state, key, plaintext_b64, &plaintext_bytes);
-
-        Ok(AwsResponse::json(
-            StatusCode::OK,
-            serde_json::to_string(&json!({
-                "CiphertextBlob": ciphertext_b64,
-                "KeyId": key.arn,
-                "EncryptionAlgorithm": "SYMMETRIC_DEFAULT",
-            }))
-            .unwrap(),
-        ))
-    }
-
-    fn decrypt(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let ciphertext_b64 = body["CiphertextBlob"].as_str().ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                "CiphertextBlob is required",
-            )
-        })?;
-
-        let accounts = self.state.read();
-        let empty = KmsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        let decoded = decode_ciphertext_envelope(state, ciphertext_b64)?;
-
-        Ok(AwsResponse::json(
-            StatusCode::OK,
-            serde_json::to_string(&json!({
-                "Plaintext": decoded.plaintext_b64,
-                "KeyId": decoded.source_arn,
-                "EncryptionAlgorithm": "SYMMETRIC_DEFAULT",
-            }))
-            .unwrap(),
-        ))
-    }
-
-    fn re_encrypt(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let ciphertext_b64 = body["CiphertextBlob"].as_str().ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                "CiphertextBlob is required",
-            )
-        })?;
-        let dest_key_id = body["DestinationKeyId"].as_str().ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                "DestinationKeyId is required",
-            )
-        })?;
-
-        let accounts = self.state.read();
-        let empty = KmsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        let decoded = decode_ciphertext_envelope(state, ciphertext_b64)?;
-
-        let dest_resolved =
-            Self::resolve_key_id_with_state(state, dest_key_id).ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "NotFoundException",
-                    format!("Key '{dest_key_id}' does not exist"),
-                )
-            })?;
-
-        let dest_key = state.keys.get(&dest_resolved).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "KMSInternalException",
-                "Key state became inconsistent",
-            )
-        })?;
-
-        let plaintext_bytes = base64::engine::general_purpose::STANDARD
-            .decode(&decoded.plaintext_b64)
-            .unwrap_or_default();
-        let new_ciphertext_b64 = if let Some(ref material) = dest_key.imported_material_bytes {
-            // Imported-key path: keep the legacy XOR envelope so consumers
-            // that already round-trip via key material can still decrypt.
-            let xored: Vec<u8> = plaintext_bytes
-                .iter()
-                .enumerate()
-                .map(|(i, b)| b ^ material[i % material.len()])
-                .collect();
-            let xored_b64 = base64::engine::general_purpose::STANDARD.encode(&xored);
-            let envelope = format!("fakecloud-imported:{}:{xored_b64}", dest_key.key_id);
-            base64::engine::general_purpose::STANDARD.encode(envelope.as_bytes())
-        } else {
-            // Default path: wrap the recovered plaintext under the
-            // destination key with the AWS-shaped binary blob.
-            let blob =
-                crate::blob::encode(&state.master_key_bytes, &dest_key.key_id, &plaintext_bytes);
-            base64::engine::general_purpose::STANDARD.encode(&blob)
-        };
-
-        Ok(AwsResponse::json(
-            StatusCode::OK,
-            serde_json::to_string(&json!({
-                "CiphertextBlob": new_ciphertext_b64,
-                "KeyId": dest_key.arn,
-                "SourceKeyId": decoded.source_arn,
-                "SourceEncryptionAlgorithm": "SYMMETRIC_DEFAULT",
-                "DestinationEncryptionAlgorithm": "SYMMETRIC_DEFAULT",
-            }))
-            .unwrap(),
-        ))
-    }
-
-    fn generate_data_key(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let key_id = Self::require_key_id(&body)?;
-
-        let resolved = self
-            .resolve_key_id_for(&req.account_id, &req.region, &key_id)
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "NotFoundException",
-                    format!("Key '{key_id}' does not exist"),
-                )
-            })?;
-
-        let accounts = self.state.read();
-        let empty = KmsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        let key = state.keys.get(&resolved).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "KMSInternalException",
-                "Key state became inconsistent",
-            )
-        })?;
-        if !key.enabled {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "DisabledException",
-                format!("Key '{}' is disabled", key.arn),
-            ));
-        }
-
-        let num_bytes = data_key_size_from_body(&body)?;
-
-        let data_key_bytes: Vec<u8> = rand_bytes(num_bytes);
-        let plaintext_b64 = base64::engine::general_purpose::STANDARD.encode(&data_key_bytes);
-
-        // Wrap the data key in the AWS-shaped binary blob.
-        let blob = crate::blob::encode(&state.master_key_bytes, &key.key_id, &data_key_bytes);
-        let ciphertext_b64 = base64::engine::general_purpose::STANDARD.encode(&blob);
-
-        Ok(AwsResponse::json(
-            StatusCode::OK,
-            serde_json::to_string(&json!({
-                "Plaintext": plaintext_b64,
-                "CiphertextBlob": ciphertext_b64,
-                "KeyId": key.arn,
-            }))
-            .unwrap(),
-        ))
-    }
-
-    fn generate_data_key_without_plaintext(
-        &self,
-        req: &AwsRequest,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let key_id = Self::require_key_id(&body)?;
-
-        let resolved = self
-            .resolve_key_id_for(&req.account_id, &req.region, &key_id)
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "NotFoundException",
-                    format!("Key '{key_id}' does not exist"),
-                )
-            })?;
-
-        let accounts = self.state.read();
-        let empty = KmsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        let key = state.keys.get(&resolved).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "KMSInternalException",
-                "Key state became inconsistent",
-            )
-        })?;
-        if !key.enabled {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "DisabledException",
-                format!("Key '{}' is disabled", key.arn),
-            ));
-        }
-
-        let num_bytes = data_key_size_from_body(&body)?;
-        let data_key_bytes: Vec<u8> = rand_bytes(num_bytes);
-        let _ = base64::engine::general_purpose::STANDARD.encode(&data_key_bytes);
-        let blob = crate::blob::encode(&state.master_key_bytes, &key.key_id, &data_key_bytes);
-        let ciphertext_b64 = base64::engine::general_purpose::STANDARD.encode(&blob);
-
-        Ok(AwsResponse::json(
-            StatusCode::OK,
-            serde_json::to_string(&json!({
-                "CiphertextBlob": ciphertext_b64,
-                "KeyId": key.arn,
-            }))
-            .unwrap(),
-        ))
-    }
-
-    fn generate_random(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-
-        // CustomKeyStoreId is accepted for API compatibility but has no effect on
-        // random number generation in this emulator.
-        validate_optional_string_length(
-            "customKeyStoreId",
-            body["CustomKeyStoreId"].as_str(),
-            1,
-            64,
-        )?;
-
-        let num_bytes = body["NumberOfBytes"].as_u64().unwrap_or(32) as usize;
-
-        validate_range_i64("numberOfBytes", num_bytes as i64, 1, 1024)?;
-
-        let random_bytes = rand_bytes(num_bytes);
-        let b64 = base64::engine::general_purpose::STANDARD.encode(&random_bytes);
-
-        Ok(AwsResponse::json(
-            StatusCode::OK,
-            serde_json::to_string(&json!({
-                "Plaintext": b64,
-            }))
-            .unwrap(),
-        ))
-    }
-
-    fn create_alias(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let alias_name = require_string_field(&body, "AliasName")?;
-        let target_key_id = require_string_field(&body, "TargetKeyId")?;
-
-        validate_alias_name(&alias_name)?;
-        validate_alias_target(&target_key_id)?;
-
-        let resolved = self
-            .resolve_key_id_for(&req.account_id, &req.region, &target_key_id)
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "NotFoundException",
-                    format!("Key '{target_key_id}' does not exist"),
-                )
-            })?;
-
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-
-        if state.aliases.contains_key(&alias_name) {
-            let alias_arn = format!(
-                "arn:aws:kms:{}:{}:{}",
-                state.region, state.account_id, alias_name
-            );
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "AlreadyExistsException",
-                format!("An alias with the name {alias_arn} already exists"),
-            ));
-        }
-
-        let alias_arn = format!(
-            "arn:aws:kms:{}:{}:{}",
-            state.region, state.account_id, alias_name
-        );
-
-        state.aliases.insert(
-            alias_name.clone(),
-            KmsAlias {
-                alias_name,
-                alias_arn,
-                target_key_id: resolved,
-                creation_date: Utc::now().timestamp() as f64,
-            },
-        );
-
-        Ok(AwsResponse::json(StatusCode::OK, "{}"))
-    }
-
-    fn delete_alias(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let alias_name = body["AliasName"].as_str().ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                "AliasName is required",
-            )
-        })?;
-
-        if !alias_name.starts_with("alias/") {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                "Invalid identifier",
-            ));
-        }
-
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-        if state.aliases.remove(alias_name).is_none() {
-            let alias_arn = format!(
-                "arn:aws:kms:{}:{}:{}",
-                state.region, state.account_id, alias_name
-            );
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "NotFoundException",
-                format!("Alias {alias_arn} is not found."),
-            ));
-        }
-
-        Ok(AwsResponse::json(StatusCode::OK, "{}"))
-    }
-
-    fn update_alias(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let alias_name = body["AliasName"].as_str().ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                "AliasName is required",
-            )
-        })?;
-        let target_key_id = body["TargetKeyId"].as_str().ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                "TargetKeyId is required",
-            )
-        })?;
-
-        let resolved = self
-            .resolve_key_id_for(&req.account_id, &req.region, target_key_id)
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "NotFoundException",
-                    format!("Key '{target_key_id}' does not exist"),
-                )
-            })?;
-
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-        let alias = state.aliases.get_mut(alias_name).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "NotFoundException",
-                format!("Alias '{alias_name}' does not exist"),
-            )
-        })?;
-
-        alias.target_key_id = resolved;
-
-        Ok(AwsResponse::json(StatusCode::OK, "{}"))
-    }
-
-    fn list_aliases(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-
-        validate_optional_json_range("limit", &body["Limit"], 1, 100)?;
-        validate_optional_string_length("marker", body["Marker"].as_str(), 1, 320)?;
-
-        if !body["KeyId"].is_null() && !body["KeyId"].is_string() {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                "KeyId must be a string",
-            ));
-        }
-        validate_optional_string_length("keyId", body["KeyId"].as_str(), 1, 2048)?;
-
-        let key_id_filter = body["KeyId"].as_str();
-
-        let accounts = self.state.read();
-        let empty = KmsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-
-        // Resolve key_id_filter to actual key ID if needed
-        let resolved_filter =
-            key_id_filter.and_then(|kid| Self::resolve_key_id_with_state(state, kid));
-
-        let aliases: Vec<Value> = state
-            .aliases
-            .values()
-            .filter(|a| match (&resolved_filter, key_id_filter) {
-                (Some(r), _) => a.target_key_id == *r,
-                (None, Some(_)) => false,
-                (None, None) => true,
-            })
-            .map(|a| {
-                json!({
-                    "AliasName": a.alias_name,
-                    "AliasArn": a.alias_arn,
-                    "TargetKeyId": a.target_key_id,
-                })
-            })
-            .collect();
-
-        Ok(AwsResponse::json(
-            StatusCode::OK,
-            serde_json::to_string(&json!({
-                "Aliases": aliases,
-                "Truncated": false,
-            }))
-            .unwrap(),
-        ))
-    }
-
     fn tag_resource(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
         let body = req.json_body();
         let key_id = Self::require_key_id(&body)?;
@@ -1611,565 +1147,6 @@ impl KmsService {
         ))
     }
 
-    fn sign(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let key_id = Self::require_key_id(&body)?;
-        let message_b64 = body["Message"].as_str().unwrap_or("");
-        let signing_algorithm = body["SigningAlgorithm"].as_str().unwrap_or("");
-
-        // Validate message
-        let message_bytes = base64::engine::general_purpose::STANDARD
-            .decode(message_b64)
-            .unwrap_or_default();
-
-        if message_bytes.is_empty() {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                "1 validation error detected: Value at 'Message' failed to satisfy constraint: Member must have length greater than or equal to 1",
-            ));
-        }
-
-        let resolved = self
-            .resolve_key_id_for(&req.account_id, &req.region, &key_id)
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "NotFoundException",
-                    format!("Key '{key_id}' does not exist"),
-                )
-            })?;
-
-        let accounts = self.state.read();
-        let empty = KmsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        let key = state.keys.get(&resolved).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "KMSInternalException",
-                "Key state became inconsistent",
-            )
-        })?;
-
-        // Validate key usage
-        if key.key_usage != "SIGN_VERIFY" {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                format!(
-                    "1 validation error detected: Value '{}' at 'KeyId' failed to satisfy constraint: Member must point to a key with usage: 'SIGN_VERIFY'",
-                    resolved
-                ),
-            ));
-        }
-
-        // Validate signing algorithm against key's supported algorithms
-        let valid_algs = key.signing_algorithms.as_deref().unwrap_or(&[]);
-        if !valid_algs.iter().any(|a| a == signing_algorithm) {
-            let set: Vec<String> = if valid_algs.is_empty() {
-                VALID_SIGNING_ALGORITHMS
-                    .iter()
-                    .map(|s| s.to_string())
-                    .collect()
-            } else {
-                valid_algs.to_vec()
-            };
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                format!(
-                    "1 validation error detected: Value '{}' at 'SigningAlgorithm' failed to satisfy constraint: Member must satisfy enum value set: {}",
-                    signing_algorithm, fmt_enum_set(&set)
-                ),
-            ));
-        }
-
-        // Generate a fake signature
-        let sig_data = format!(
-            "fakecloud-sig:{}:{}:{}",
-            key.key_id, signing_algorithm, message_b64
-        );
-        let signature_b64 = base64::engine::general_purpose::STANDARD.encode(sig_data.as_bytes());
-
-        Ok(AwsResponse::json(
-            StatusCode::OK,
-            serde_json::to_string(&json!({
-                "Signature": signature_b64,
-                "SigningAlgorithm": signing_algorithm,
-                "KeyId": key.arn,
-            }))
-            .unwrap(),
-        ))
-    }
-
-    fn verify(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let key_id = Self::require_key_id(&body)?;
-        let message_b64 = body["Message"].as_str().unwrap_or("");
-        let signature_b64 = body["Signature"].as_str().unwrap_or("");
-        let signing_algorithm = body["SigningAlgorithm"].as_str().unwrap_or("");
-
-        require_non_empty_b64("Message", message_b64)?;
-        require_non_empty_b64("Signature", signature_b64)?;
-
-        let resolved = self
-            .resolve_key_id_for(&req.account_id, &req.region, &key_id)
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "NotFoundException",
-                    format!("Key '{key_id}' does not exist"),
-                )
-            })?;
-
-        let accounts = self.state.read();
-        let empty = KmsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        let key = state.keys.get(&resolved).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "KMSInternalException",
-                "Key state became inconsistent",
-            )
-        })?;
-
-        validate_key_usage_signing(key, &resolved)?;
-        validate_signing_algorithm(key, signing_algorithm)?;
-
-        // Check if signature matches the deterministic fakecloud signature.
-        let expected_sig_data = format!(
-            "fakecloud-sig:{}:{}:{}",
-            key.key_id, signing_algorithm, message_b64
-        );
-        let expected_signature_b64 =
-            base64::engine::general_purpose::STANDARD.encode(expected_sig_data.as_bytes());
-
-        let signature_valid = signature_b64 == expected_signature_b64;
-
-        Ok(AwsResponse::json(
-            StatusCode::OK,
-            serde_json::to_string(&json!({
-                "SignatureValid": signature_valid,
-                "SigningAlgorithm": signing_algorithm,
-                "KeyId": key.arn,
-            }))
-            .unwrap(),
-        ))
-    }
-
-    fn get_public_key(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let key_id = Self::require_key_id(&body)?;
-
-        let resolved = self
-            .resolve_key_id_for(&req.account_id, &req.region, &key_id)
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "NotFoundException",
-                    format!("Key '{key_id}' does not exist"),
-                )
-            })?;
-
-        let accounts = self.state.read();
-        let empty = KmsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        let key = state.keys.get(&resolved).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "KMSInternalException",
-                "Key state became inconsistent",
-            )
-        })?;
-
-        // Generate a fake DER-encoded public key
-        let fake_public_key = generate_fake_public_key(&key.key_spec);
-        let public_key_b64 = base64::engine::general_purpose::STANDARD.encode(&fake_public_key);
-
-        let mut response = json!({
-            "KeyId": key.arn,
-            "KeySpec": key.key_spec,
-            "KeyUsage": key.key_usage,
-            "PublicKey": public_key_b64,
-            "CustomerMasterKeySpec": key.key_spec,
-        });
-
-        if let Some(ref signing_algs) = key.signing_algorithms {
-            response["SigningAlgorithms"] = json!(signing_algs);
-        }
-        if let Some(ref enc_algs) = key.encryption_algorithms {
-            response["EncryptionAlgorithms"] = json!(enc_algs);
-        }
-
-        Ok(AwsResponse::json(
-            StatusCode::OK,
-            serde_json::to_string(&response).unwrap(),
-        ))
-    }
-
-    fn create_grant(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let key_id = Self::require_key_id(&body)?;
-
-        let resolved = self
-            .resolve_key_id_for(&req.account_id, &req.region, &key_id)
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "NotFoundException",
-                    format!("Key '{key_id}' does not exist"),
-                )
-            })?;
-
-        let grantee_principal = body["GranteePrincipal"].as_str().unwrap_or("").to_string();
-        let retiring_principal = body["RetiringPrincipal"].as_str().map(|s| s.to_string());
-        let operations: Vec<String> = body["Operations"]
-            .as_array()
-            .map(|arr| {
-                arr.iter()
-                    .filter_map(|v| v.as_str().map(|s| s.to_string()))
-                    .collect()
-            })
-            .unwrap_or_default();
-        let constraints = if body["Constraints"].is_null() {
-            None
-        } else {
-            Some(body["Constraints"].clone())
-        };
-        let name = body["Name"].as_str().map(|s| s.to_string());
-
-        let grant_id = Uuid::new_v4().to_string();
-        let grant_token = Uuid::new_v4().to_string();
-
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-        state.grants.push(KmsGrant {
-            grant_id: grant_id.clone(),
-            grant_token: grant_token.clone(),
-            key_id: resolved,
-            grantee_principal,
-            retiring_principal,
-            operations,
-            constraints,
-            name,
-            creation_date: Utc::now().timestamp() as f64,
-        });
-
-        Ok(AwsResponse::json(
-            StatusCode::OK,
-            serde_json::to_string(&json!({
-                "GrantId": grant_id,
-                "GrantToken": grant_token,
-            }))
-            .unwrap(),
-        ))
-    }
-
-    fn list_grants(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let key_id = Self::require_key_id(&body)?;
-
-        let resolved = self
-            .resolve_key_id_for(&req.account_id, &req.region, &key_id)
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "NotFoundException",
-                    format!("Key '{key_id}' does not exist"),
-                )
-            })?;
-
-        let grant_id_filter = body["GrantId"].as_str();
-
-        let accounts = self.state.read();
-        let empty = KmsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        let grants: Vec<Value> = state
-            .grants
-            .iter()
-            .filter(|g| g.key_id == resolved)
-            .filter(|g| {
-                if let Some(gid) = grant_id_filter {
-                    g.grant_id == gid
-                } else {
-                    true
-                }
-            })
-            .map(|g| grant_to_json(g, &req.account_id))
-            .collect();
-
-        Ok(AwsResponse::json(
-            StatusCode::OK,
-            serde_json::to_string(&json!({
-                "Grants": grants,
-                "Truncated": false,
-            }))
-            .unwrap(),
-        ))
-    }
-
-    fn list_retirable_grants(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-
-        validate_required("RetiringPrincipal", &body["RetiringPrincipal"])?;
-        let retiring_principal = body["RetiringPrincipal"].as_str().ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                "RetiringPrincipal must be a string",
-            )
-        })?;
-        validate_string_length("retiringPrincipal", retiring_principal, 1, 256)?;
-        validate_optional_json_range("limit", &body["Limit"], 1, 1000)?;
-        validate_optional_string_length("marker", body["Marker"].as_str(), 1, 320)?;
-
-        let limit = body["Limit"].as_i64().unwrap_or(1000) as usize;
-        let marker = body["Marker"].as_str();
-
-        let accounts = self.state.read();
-        let empty = KmsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        let all_grants: Vec<Value> = state
-            .grants
-            .iter()
-            .filter(|g| {
-                g.retiring_principal
-                    .as_deref()
-                    .is_some_and(|rp| rp == retiring_principal)
-            })
-            .map(|g| grant_to_json(g, &req.account_id))
-            .collect();
-
-        let start = if let Some(m) = marker {
-            all_grants
-                .iter()
-                .position(|g| g["GrantId"].as_str() == Some(m))
-                .map(|pos| pos + 1)
-                .unwrap_or(0)
-        } else {
-            0
-        };
-
-        let page = &all_grants[start..all_grants.len().min(start + limit)];
-        let truncated = start + limit < all_grants.len();
-
-        let mut result = json!({
-            "Grants": page,
-            "Truncated": truncated,
-        });
-
-        if truncated {
-            if let Some(last) = page.last() {
-                result["NextMarker"] = last["GrantId"].clone();
-            }
-        }
-
-        Ok(AwsResponse::json(
-            StatusCode::OK,
-            serde_json::to_string(&result).unwrap(),
-        ))
-    }
-
-    fn revoke_grant(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let key_id = Self::require_key_id(&body)?;
-        let grant_id = body["GrantId"].as_str().unwrap_or("");
-
-        let resolved = self
-            .resolve_key_id_for(&req.account_id, &req.region, &key_id)
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "NotFoundException",
-                    format!("Key '{key_id}' does not exist"),
-                )
-            })?;
-
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-        let idx = state
-            .grants
-            .iter()
-            .position(|g| g.key_id == resolved && g.grant_id == grant_id);
-
-        match idx {
-            Some(i) => {
-                state.grants.remove(i);
-                Ok(AwsResponse::json(StatusCode::OK, "{}"))
-            }
-            None => Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "NotFoundException",
-                format!("Grant ID {grant_id} not found"),
-            )),
-        }
-    }
-
-    fn retire_grant(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let grant_token = body["GrantToken"].as_str();
-        let grant_id = body["GrantId"].as_str();
-        let key_id = body["KeyId"].as_str();
-
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-
-        let idx = if let Some(token) = grant_token {
-            state.grants.iter().position(|g| g.grant_token == token)
-        } else if let (Some(kid), Some(gid)) = (key_id, grant_id) {
-            let resolved = Self::resolve_key_id_with_state(state, kid);
-            resolved.and_then(|r| {
-                state
-                    .grants
-                    .iter()
-                    .position(|g| g.key_id == r && g.grant_id == gid)
-            })
-        } else {
-            None
-        };
-
-        match idx {
-            Some(i) => {
-                state.grants.remove(i);
-                Ok(AwsResponse::json(StatusCode::OK, "{}"))
-            }
-            None => Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "NotFoundException",
-                "Grant not found",
-            )),
-        }
-    }
-
-    fn generate_mac(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let key_id = Self::require_key_id(&body)?;
-        let mac_algorithm = body["MacAlgorithm"].as_str().unwrap_or("").to_string();
-        let message_b64 = body["Message"].as_str().unwrap_or("");
-
-        let resolved = self
-            .resolve_key_id_for(&req.account_id, &req.region, &key_id)
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "NotFoundException",
-                    format!("Key '{key_id}' does not exist"),
-                )
-            })?;
-
-        let accounts = self.state.read();
-        let empty = KmsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        let key = state.keys.get(&resolved).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "KMSInternalException",
-                "Key state became inconsistent",
-            )
-        })?;
-
-        // Validate key usage
-        if key.key_usage != "GENERATE_VERIFY_MAC" {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "InvalidKeyUsageException",
-                format!("Key '{}' is not a GENERATE_VERIFY_MAC key", key.arn),
-            ));
-        }
-
-        // Validate key spec supports MAC
-        if key.mac_algorithms.is_none() {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "InvalidKeyUsageException",
-                format!("Key '{}' does not support MAC operations", key.arn),
-            ));
-        }
-
-        // Generate fake MAC
-        let mac_data = format!(
-            "fakecloud-mac:{}:{}:{}",
-            key.key_id, mac_algorithm, message_b64
-        );
-        let mac_b64 = base64::engine::general_purpose::STANDARD.encode(mac_data.as_bytes());
-
-        Ok(AwsResponse::json(
-            StatusCode::OK,
-            serde_json::to_string(&json!({
-                "Mac": mac_b64,
-                "KeyId": key.key_id,
-                "MacAlgorithm": mac_algorithm,
-            }))
-            .unwrap(),
-        ))
-    }
-
-    fn verify_mac(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let key_id = Self::require_key_id(&body)?;
-        let mac_algorithm = body["MacAlgorithm"].as_str().unwrap_or("").to_string();
-        let message_b64 = body["Message"].as_str().unwrap_or("");
-        let mac_b64 = body["Mac"].as_str().unwrap_or("");
-
-        let resolved = self
-            .resolve_key_id_for(&req.account_id, &req.region, &key_id)
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "NotFoundException",
-                    format!("Key '{key_id}' does not exist"),
-                )
-            })?;
-
-        let accounts = self.state.read();
-        let empty = KmsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        let key = state.keys.get(&resolved).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "KMSInternalException",
-                "Key state became inconsistent",
-            )
-        })?;
-
-        // Validate key usage
-        if key.key_usage != "GENERATE_VERIFY_MAC" {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "InvalidKeyUsageException",
-                format!("Key '{}' is not a GENERATE_VERIFY_MAC key", key.arn),
-            ));
-        }
-
-        // Check if MAC matches
-        let expected_mac_data = format!(
-            "fakecloud-mac:{}:{}:{}",
-            key.key_id, mac_algorithm, message_b64
-        );
-        let expected_mac_b64 =
-            base64::engine::general_purpose::STANDARD.encode(expected_mac_data.as_bytes());
-
-        let mac_valid = mac_b64 == expected_mac_b64;
-
-        if !mac_valid {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "KMSInvalidMacException",
-                "MAC verification failed",
-            ));
-        }
-
-        Ok(AwsResponse::json(
-            StatusCode::OK,
-            serde_json::to_string(&json!({
-                "KeyId": key.key_id,
-                "MacAlgorithm": mac_algorithm,
-                "MacValid": true,
-            }))
-            .unwrap(),
-        ))
-    }
-
     fn replicate_key(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
         let body = req.json_body();
         let key_id = Self::require_key_id(&body)?;
@@ -2261,377 +1238,6 @@ impl KmsService {
         ))
     }
 
-    fn generate_data_key_pair(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let key_id = Self::require_key_id(&body)?;
-        let key_pair_spec = body["KeyPairSpec"]
-            .as_str()
-            .unwrap_or("RSA_2048")
-            .to_string();
-
-        let resolved = self
-            .resolve_key_id_for(&req.account_id, &req.region, &key_id)
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "NotFoundException",
-                    format!("Key '{key_id}' does not exist"),
-                )
-            })?;
-
-        let accounts = self.state.read();
-        let empty = KmsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        let key = state.keys.get(&resolved).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "KMSInternalException",
-                "Key state became inconsistent",
-            )
-        })?;
-        if !key.enabled {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "DisabledException",
-                format!("Key '{}' is disabled", key.arn),
-            ));
-        }
-
-        let public_key_bytes = generate_fake_public_key(&key_pair_spec);
-        let private_key_bytes = rand_bytes(256);
-        let public_key_b64 = base64::engine::general_purpose::STANDARD.encode(&public_key_bytes);
-        let private_plaintext_b64 =
-            base64::engine::general_purpose::STANDARD.encode(&private_key_bytes);
-
-        // Wrap the private key in the AWS-shaped binary blob.
-        let blob = crate::blob::encode(&state.master_key_bytes, &key.key_id, &private_key_bytes);
-        let private_ciphertext_b64 = base64::engine::general_purpose::STANDARD.encode(&blob);
-
-        Ok(AwsResponse::json(
-            StatusCode::OK,
-            serde_json::to_string(&json!({
-                "KeyId": key.arn,
-                "KeyPairSpec": key_pair_spec,
-                "PublicKey": public_key_b64,
-                "PrivateKeyPlaintext": private_plaintext_b64,
-                "PrivateKeyCiphertextBlob": private_ciphertext_b64,
-            }))
-            .unwrap(),
-        ))
-    }
-
-    fn generate_data_key_pair_without_plaintext(
-        &self,
-        req: &AwsRequest,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let key_id = Self::require_key_id(&body)?;
-        let key_pair_spec = body["KeyPairSpec"]
-            .as_str()
-            .unwrap_or("RSA_2048")
-            .to_string();
-
-        let resolved = self
-            .resolve_key_id_for(&req.account_id, &req.region, &key_id)
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "NotFoundException",
-                    format!("Key '{key_id}' does not exist"),
-                )
-            })?;
-
-        let accounts = self.state.read();
-        let empty = KmsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        let key = state.keys.get(&resolved).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "KMSInternalException",
-                "Key state became inconsistent",
-            )
-        })?;
-        if !key.enabled {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "DisabledException",
-                format!("Key '{}' is disabled", key.arn),
-            ));
-        }
-
-        let public_key_bytes = generate_fake_public_key(&key_pair_spec);
-        let private_key_bytes = rand_bytes(256);
-        let public_key_b64 = base64::engine::general_purpose::STANDARD.encode(&public_key_bytes);
-
-        let blob = crate::blob::encode(&state.master_key_bytes, &key.key_id, &private_key_bytes);
-        let private_ciphertext_b64 = base64::engine::general_purpose::STANDARD.encode(&blob);
-
-        Ok(AwsResponse::json(
-            StatusCode::OK,
-            serde_json::to_string(&json!({
-                "KeyId": key.arn,
-                "KeyPairSpec": key_pair_spec,
-                "PublicKey": public_key_b64,
-                "PrivateKeyCiphertextBlob": private_ciphertext_b64,
-            }))
-            .unwrap(),
-        ))
-    }
-
-    fn derive_shared_secret(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let key_id = Self::require_key_id(&body)?;
-        let _key_agreement_algorithm = body["KeyAgreementAlgorithm"]
-            .as_str()
-            .unwrap_or("ECDH")
-            .to_string();
-        let _public_key = body["PublicKey"].as_str().ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                "PublicKey is required",
-            )
-        })?;
-
-        let resolved = self
-            .resolve_key_id_for(&req.account_id, &req.region, &key_id)
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "NotFoundException",
-                    format!("Key '{key_id}' does not exist"),
-                )
-            })?;
-
-        let accounts = self.state.read();
-        let empty = KmsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        let key = state.keys.get(&resolved).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "KMSInternalException",
-                "Key state became inconsistent",
-            )
-        })?;
-
-        if !key.enabled {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "DisabledException",
-                format!("Key '{}' is disabled", key.arn),
-            ));
-        }
-
-        // Key must be asymmetric (KEY_AGREEMENT usage)
-        if key.key_usage != "KEY_AGREEMENT" {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "InvalidKeyUsageException",
-                format!(
-                    "Key '{}' usage is '{}', not KEY_AGREEMENT",
-                    key.arn, key.key_usage
-                ),
-            ));
-        }
-
-        // Deterministic shared secret: SHA-256(private_key_seed || public_key_bytes)
-        // Both parties using the correct keys will derive the same result.
-        let public_key_bytes = base64::engine::general_purpose::STANDARD
-            .decode(_public_key)
-            .unwrap_or_default();
-
-        use sha2::{Digest, Sha256};
-        let mut hasher = Sha256::new();
-        hasher.update(&key.private_key_seed);
-        hasher.update(&public_key_bytes);
-        let shared_secret_bytes = hasher.finalize();
-        let shared_secret_b64 =
-            base64::engine::general_purpose::STANDARD.encode(shared_secret_bytes);
-
-        Ok(AwsResponse::json(
-            StatusCode::OK,
-            serde_json::to_string(&json!({
-                "KeyId": key.arn,
-                "SharedSecret": shared_secret_b64,
-                "KeyAgreementAlgorithm": "ECDH",
-                "KeyOrigin": key.origin,
-            }))
-            .unwrap(),
-        ))
-    }
-
-    fn get_parameters_for_import(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let key_id = Self::require_key_id(&body)?;
-
-        let resolved = self
-            .resolve_key_id_for(&req.account_id, &req.region, &key_id)
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "NotFoundException",
-                    format!("Key '{key_id}' does not exist"),
-                )
-            })?;
-
-        let accounts = self.state.read();
-        let empty = KmsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-        let key = state.keys.get(&resolved).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "KMSInternalException",
-                "Key state became inconsistent",
-            )
-        })?;
-
-        if key.origin != "EXTERNAL" {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "UnsupportedOperationException",
-                format!("Key '{}' origin is '{}', not EXTERNAL", key.arn, key.origin),
-            ));
-        }
-
-        let import_token_bytes = rand_bytes(64);
-        let import_token_b64 =
-            base64::engine::general_purpose::STANDARD.encode(&import_token_bytes);
-        let public_key_bytes = generate_fake_public_key("RSA_2048");
-        let public_key_b64 = base64::engine::general_purpose::STANDARD.encode(&public_key_bytes);
-
-        // Valid for 24 hours
-        let parameters_valid_to = Utc::now().timestamp() as f64 + 86400.0;
-
-        Ok(AwsResponse::json(
-            StatusCode::OK,
-            serde_json::to_string(&json!({
-                "KeyId": key.arn,
-                "ImportToken": import_token_b64,
-                "PublicKey": public_key_b64,
-                "ParametersValidTo": parameters_valid_to,
-            }))
-            .unwrap(),
-        ))
-    }
-
-    fn import_key_material(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let key_id = Self::require_key_id(&body)?;
-
-        let _import_token = body["ImportToken"].as_str().ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                "ImportToken is required",
-            )
-        })?;
-
-        let encrypted_key_material = body["EncryptedKeyMaterial"].as_str().ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                "EncryptedKeyMaterial is required",
-            )
-        })?;
-
-        let resolved = self
-            .resolve_key_id_for(&req.account_id, &req.region, &key_id)
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "NotFoundException",
-                    format!("Key '{key_id}' does not exist"),
-                )
-            })?;
-
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-        let key = state.keys.get_mut(&resolved).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "NotFoundException",
-                format!("Key '{key_id}' does not exist"),
-            )
-        })?;
-
-        if key.origin != "EXTERNAL" {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "UnsupportedOperationException",
-                format!("Key '{}' origin is '{}', not EXTERNAL", key.arn, key.origin),
-            ));
-        }
-
-        // Store the imported material bytes for use in encrypt/decrypt.
-        // In real AWS, the material is unwrapped with the import RSA key.
-        // Here we treat the EncryptedKeyMaterial as the raw key (base64-decoded).
-        let material_bytes = base64::engine::general_purpose::STANDARD
-            .decode(encrypted_key_material)
-            .map_err(|_| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "ValidationException",
-                    "EncryptedKeyMaterial is not valid base64",
-                )
-            })?;
-        if material_bytes.is_empty() {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "ValidationException",
-                "EncryptedKeyMaterial must not be empty",
-            ));
-        }
-        key.imported_key_material = true;
-        key.imported_material_bytes = Some(material_bytes);
-        key.enabled = true;
-        key.key_state = "Enabled".to_string();
-
-        Ok(AwsResponse::json(StatusCode::OK, "{}"))
-    }
-
-    fn delete_imported_key_material(
-        &self,
-        req: &AwsRequest,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        let key_id = Self::require_key_id(&body)?;
-
-        let resolved = self
-            .resolve_key_id_for(&req.account_id, &req.region, &key_id)
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "NotFoundException",
-                    format!("Key '{key_id}' does not exist"),
-                )
-            })?;
-
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-        let key = state.keys.get_mut(&resolved).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "NotFoundException",
-                format!("Key '{key_id}' does not exist"),
-            )
-        })?;
-
-        if key.origin != "EXTERNAL" {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "UnsupportedOperationException",
-                format!("Key '{}' origin is '{}', not EXTERNAL", key.arn, key.origin),
-            ));
-        }
-
-        key.imported_key_material = false;
-        key.imported_material_bytes = None;
-        key.enabled = false;
-        key.key_state = "PendingImport".to_string();
-
-        Ok(AwsResponse::json(StatusCode::OK, "{}"))
-    }
-
     fn update_primary_region(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
         let body = req.json_body();
         let key_id = Self::require_key_id(&body)?;
@@ -2683,321 +1289,16 @@ impl KmsService {
 
         Ok(AwsResponse::json(StatusCode::OK, "{}"))
     }
-
-    fn create_custom_key_store(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-
-        let name = body["CustomKeyStoreName"]
-            .as_str()
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "ValidationException",
-                    "CustomKeyStoreName is required",
-                )
-            })?
-            .to_string();
-
-        validate_string_length("customKeyStoreName", &name, 1, 256)?;
-
-        let store_type = body["CustomKeyStoreType"]
-            .as_str()
-            .unwrap_or("AWS_CLOUDHSM")
-            .to_string();
-
-        validate_optional_enum(
-            "customKeyStoreType",
-            Some(store_type.as_str()),
-            &["AWS_CLOUDHSM", "EXTERNAL_KEY_STORE"],
-        )?;
-
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-
-        // Name must be unique
-        if state
-            .custom_key_stores
-            .values()
-            .any(|s| s.custom_key_store_name == name)
-        {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "CustomKeyStoreNameInUseException",
-                format!("Custom key store name '{name}' is already in use"),
-            ));
-        }
-
-        let store_id = format!("cks-{}", Uuid::new_v4().as_simple());
-        let now = Utc::now().timestamp() as f64;
-
-        let store = CustomKeyStore {
-            custom_key_store_id: store_id.clone(),
-            custom_key_store_name: name,
-            custom_key_store_type: store_type,
-            cloud_hsm_cluster_id: body["CloudHsmClusterId"].as_str().map(|s| s.to_string()),
-            trust_anchor_certificate: body["TrustAnchorCertificate"]
-                .as_str()
-                .map(|s| s.to_string()),
-            connection_state: "DISCONNECTED".to_string(),
-            creation_date: now,
-            xks_proxy_uri_endpoint: body["XksProxyUriEndpoint"].as_str().map(|s| s.to_string()),
-            xks_proxy_uri_path: body["XksProxyUriPath"].as_str().map(|s| s.to_string()),
-            xks_proxy_vpc_endpoint_service_name: body["XksProxyVpcEndpointServiceName"]
-                .as_str()
-                .map(|s| s.to_string()),
-            xks_proxy_connectivity: body["XksProxyConnectivity"].as_str().map(|s| s.to_string()),
-        };
-
-        state.custom_key_stores.insert(store_id.clone(), store);
-
-        Ok(AwsResponse::json(
-            StatusCode::OK,
-            serde_json::to_string(&json!({ "CustomKeyStoreId": store_id })).unwrap(),
-        ))
-    }
-
-    fn delete_custom_key_store(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-
-        let store_id = body["CustomKeyStoreId"]
-            .as_str()
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "ValidationException",
-                    "CustomKeyStoreId is required",
-                )
-            })?
-            .to_string();
-
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-
-        let store = state.custom_key_stores.get(&store_id).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "CustomKeyStoreNotFoundException",
-                format!("Custom key store '{store_id}' does not exist"),
-            )
-        })?;
-
-        if store.connection_state == "CONNECTED" {
-            return Err(AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "CustomKeyStoreHasCMKsException",
-                "Cannot delete a connected custom key store. Disconnect it first.",
-            ));
-        }
-
-        state.custom_key_stores.remove(&store_id);
-
-        Ok(AwsResponse::json(StatusCode::OK, "{}"))
-    }
-
-    fn describe_custom_key_stores(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-        validate_optional_string_length(
-            "customKeyStoreName",
-            body["CustomKeyStoreName"].as_str(),
-            1,
-            256,
-        )?;
-        validate_optional_json_range("limit", &body["Limit"], 1, 1000)?;
-        validate_optional_string_length("marker", body["Marker"].as_str(), 1, 1024)?;
-
-        let filter_id = body["CustomKeyStoreId"].as_str();
-        let filter_name = body["CustomKeyStoreName"].as_str();
-        let limit = body["Limit"].as_i64().unwrap_or(1000) as usize;
-        let marker = body["Marker"].as_str();
-
-        let accounts = self.state.read();
-        let empty = KmsState::new(&req.account_id, &req.region);
-        let state = accounts.get(&req.account_id).unwrap_or(&empty);
-
-        let mut stores: Vec<&CustomKeyStore> = state
-            .custom_key_stores
-            .values()
-            .filter(|s| {
-                if let Some(id) = filter_id {
-                    return s.custom_key_store_id == id;
-                }
-                if let Some(name) = filter_name {
-                    return s.custom_key_store_name == name;
-                }
-                true
-            })
-            .collect();
-
-        stores.sort_by(|a, b| a.custom_key_store_id.cmp(&b.custom_key_store_id));
-
-        // If filtering by ID and not found, return error
-        if let Some(id) = filter_id {
-            if stores.is_empty() {
-                return Err(AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "CustomKeyStoreNotFoundException",
-                    format!("Custom key store '{id}' does not exist"),
-                ));
-            }
-        }
-
-        let start = marker
-            .and_then(|m| {
-                stores
-                    .iter()
-                    .position(|s| s.custom_key_store_id == m)
-                    .map(|p| p + 1)
-            })
-            .unwrap_or(0);
-
-        let page: Vec<_> = stores.iter().skip(start).take(limit).collect();
-        let truncated = start + page.len() < stores.len();
-
-        let entries: Vec<Value> = page.iter().map(|s| custom_key_store_json(s)).collect();
-
-        let mut resp = json!({ "CustomKeyStores": entries, "Truncated": truncated });
-        if truncated {
-            if let Some(last) = page.last() {
-                resp["NextMarker"] = json!(last.custom_key_store_id);
-            }
-        }
-
-        Ok(AwsResponse::json(
-            StatusCode::OK,
-            serde_json::to_string(&resp).unwrap(),
-        ))
-    }
-
-    fn connect_custom_key_store(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-
-        let store_id = body["CustomKeyStoreId"]
-            .as_str()
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "ValidationException",
-                    "CustomKeyStoreId is required",
-                )
-            })?
-            .to_string();
-
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-
-        let store = state.custom_key_stores.get_mut(&store_id).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "CustomKeyStoreNotFoundException",
-                format!("Custom key store '{store_id}' does not exist"),
-            )
-        })?;
-
-        store.connection_state = "CONNECTED".to_string();
-
-        Ok(AwsResponse::json(StatusCode::OK, "{}"))
-    }
-
-    fn disconnect_custom_key_store(
-        &self,
-        req: &AwsRequest,
-    ) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-
-        let store_id = body["CustomKeyStoreId"]
-            .as_str()
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "ValidationException",
-                    "CustomKeyStoreId is required",
-                )
-            })?
-            .to_string();
-
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-
-        let store = state.custom_key_stores.get_mut(&store_id).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "CustomKeyStoreNotFoundException",
-                format!("Custom key store '{store_id}' does not exist"),
-            )
-        })?;
-
-        store.connection_state = "DISCONNECTED".to_string();
-
-        Ok(AwsResponse::json(StatusCode::OK, "{}"))
-    }
-
-    fn update_custom_key_store(&self, req: &AwsRequest) -> Result<AwsResponse, AwsServiceError> {
-        let body = req.json_body();
-
-        let store_id = body["CustomKeyStoreId"]
-            .as_str()
-            .ok_or_else(|| {
-                AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "ValidationException",
-                    "CustomKeyStoreId is required",
-                )
-            })?
-            .to_string();
-
-        let mut accounts = self.state.write();
-        let state = accounts.get_or_create(&req.account_id);
-
-        // Check uniqueness of new name before borrowing store mutably
-        if let Some(new_name) = body["NewCustomKeyStoreName"].as_str() {
-            if state
-                .custom_key_stores
-                .values()
-                .any(|s| s.custom_key_store_name == new_name && s.custom_key_store_id != store_id)
-            {
-                return Err(AwsServiceError::aws_error(
-                    StatusCode::BAD_REQUEST,
-                    "CustomKeyStoreNameInUseException",
-                    format!("Custom key store name '{new_name}' is already in use"),
-                ));
-            }
-        }
-
-        let store = state.custom_key_stores.get_mut(&store_id).ok_or_else(|| {
-            AwsServiceError::aws_error(
-                StatusCode::BAD_REQUEST,
-                "CustomKeyStoreNotFoundException",
-                format!("Custom key store '{store_id}' does not exist"),
-            )
-        })?;
-
-        if let Some(new_name) = body["NewCustomKeyStoreName"].as_str() {
-            store.custom_key_store_name = new_name.to_string();
-        }
-        if let Some(v) = body["CloudHsmClusterId"].as_str() {
-            store.cloud_hsm_cluster_id = Some(v.to_string());
-        }
-        if let Some(v) = body["KeyStorePassword"].as_str() {
-            // In a real implementation this would update the password;
-            // we just accept it silently.
-            let _ = v;
-        }
-        if let Some(v) = body["XksProxyUriEndpoint"].as_str() {
-            store.xks_proxy_uri_endpoint = Some(v.to_string());
-        }
-        if let Some(v) = body["XksProxyUriPath"].as_str() {
-            store.xks_proxy_uri_path = Some(v.to_string());
-        }
-        if let Some(v) = body["XksProxyVpcEndpointServiceName"].as_str() {
-            store.xks_proxy_vpc_endpoint_service_name = Some(v.to_string());
-        }
-        if let Some(v) = body["XksProxyConnectivity"].as_str() {
-            store.xks_proxy_connectivity = Some(v.to_string());
-        }
-
-        Ok(AwsResponse::json(StatusCode::OK, "{}"))
-    }
 }
+
+#[path = "service_aliases.rs"]
+mod service_aliases;
+#[path = "service_crypto.rs"]
+mod service_crypto;
+#[path = "service_custom_store.rs"]
+mod service_custom_store;
+#[path = "service_grants.rs"]
+mod service_grants;
 
 #[path = "helpers.rs"]
 mod helpers;
